+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - bump required `py-meta-utils` to v0.6.1
 - `ColumnMetaOption.check_value` should raise `TypeError` or `ValueError`, not `AssertionError`
 - Move declaration of factory meta options from `ModelMetaOptionsFactory._get_meta_options()` to `ModelMetaOptionsFactory._options`
+- bugfix: use correct foreign key for the primary key on joined polymorphic models
 
 ## v0.4.0 (2018/10/09)